refactor: all read rolePermission
