+++ src/app/admin/permissions/page.tsx
@@ -1,8 +1,8 @@
 import RoleView from './RoleView'
-import { readRoles } from '@/actions/rolePermissions/read'
+import { readRolesAction } from '@/actions/rolePermissions/read'
 
 export default async function Permissions() {
-    const res = await readRoles()
+    const res = await readRolesAction()
 
     if (!res.success) throw Error(res.error ? res.error[0].message : 'error')
